update modpath and fix for #72 and #71
