--- nlmod/mfpackages/__init__.py
@@ -1,2 +1,0 @@
-from . import constant_head, recharge, surface_water, horizontal_flow_barrier, modpath
-from .gwf import *
