+++ tests/test_001_model.py
@@ -122,10 +122,20 @@ def test_create_small_model_grid_only(tmpdir, model_name="test"):
         transient_timesteps=10,
     )
 
-    _, gwf = nlmod.mfpackages.sim_tdis_gwf_ims_from_model_ds(model_ds)
+    # create simulation 
+    sim = nlmod.gwf.sim_from_model_ds(model_ds)
+    
+    # create time discretisation
+    tdis = nlmod.gwf.tdis_from_model_ds(model_ds, sim)
+    
+    # create groundwater flow model
+    gwf = nlmod.gwf.gwf_from_model_ds(model_ds, sim)
+    
+    # create ims
+    ims = nlmod.gwf.ims_to_sim(sim)
 
     # Create discretization
-    nlmod.mfpackages.dis_from_model_ds(model_ds, gwf)
+    nlmod.gwf.dis_from_model_ds(model_ds, gwf)
 
     # save model_ds
     model_ds.to_netcdf(os.path.join(tst_model_dir, "small_model.nc"))
@@ -185,43 +195,53 @@ def test_create_sea_model(tmpdir):
     model_ds = xr.open_dataset(
         os.path.join(tst_model_dir, "basic_sea_model.nc"), mask_and_scale=False
     )
-    # create modflow packages
-    _, gwf = nlmod.mfpackages.sim_tdis_gwf_ims_from_model_ds(model_ds)
+    # create simulation 
+    sim = nlmod.gwf.sim_from_model_ds(model_ds)
+    
+    # create time discretisation
+    tdis = nlmod.gwf.tdis_from_model_ds(model_ds, sim)
+    
+    # create groundwater flow model
+    gwf = nlmod.gwf.gwf_from_model_ds(model_ds, sim)
+    
+    # create ims
+    ims = nlmod.gwf.ims_to_sim(sim)
+    
     # Create discretization
-    nlmod.mfpackages.dis_from_model_ds(model_ds, gwf)
+    nlmod.gwf.dis_from_model_ds(model_ds, gwf)
 
     # create node property flow
-    nlmod.mfpackages.npf_from_model_ds(model_ds, gwf)
+    nlmod.gwf.npf_from_model_ds(model_ds, gwf)
 
     # Create the initial conditions package
-    nlmod.mfpackages.ic_from_model_ds(model_ds, gwf, starting_head=1.0)
+    nlmod.gwf.ic_from_model_ds(model_ds, gwf, starting_head=1.0)
 
     # Create the output control package
-    nlmod.mfpackages.oc_from_model_ds(model_ds, gwf)
+    nlmod.gwf.oc_from_model_ds(model_ds, gwf)
 
     # voeg grote oppervlaktewaterlichamen toe
     da_name = "surface_water"
     model_ds.update(nlmod.read.rws.get_surface_water(model_ds, da_name))
-    nlmod.mfpackages.ghb_from_model_ds(model_ds, gwf, da_name)
+    nlmod.gwf.ghb_from_model_ds(model_ds, gwf, da_name)
 
     # surface level drain
     model_ds.update(nlmod.read.ahn.get_ahn(model_ds))
-    nlmod.mfpackages.surface_drain_from_model_ds(model_ds, gwf)
+    nlmod.gwf.surface_drain_from_model_ds(model_ds, gwf)
 
     # add constant head cells at model boundaries
     model_ds.update(
-        nlmod.mfpackages.constant_head.get_chd_at_model_edge(
+        nlmod.gwf.constant_head.get_chd_at_model_edge(
             model_ds, model_ds["idomain"]
         )
     )
-    nlmod.mfpackages.chd_from_model_ds(model_ds, gwf, head="starting_head")
+    nlmod.gwf.chd_from_model_ds(model_ds, gwf, head="starting_head")
 
     # add knmi recharge to the model datasets
     model_ds.update(nlmod.read.knmi.get_recharge(model_ds))
     # create recharge package
-    nlmod.mfpackages.rch_from_model_ds(model_ds, gwf)
+    nlmod.gwf.rch_from_model_ds(model_ds, gwf)
 
-    nlmod.util.write_and_run_model(gwf, model_ds)
+    nlmod.gwf.write_and_run_model(gwf, model_ds)
 
     # gwf.simulation.write_simulation()
 
@@ -259,43 +279,53 @@ def test_create_sea_model_perlen_list(tmpdir):
     model_ds = model_ds.drop_dims("time")
     model_ds.update(new_model_ds)
 
-    # create modflow packages
-    sim, gwf = nlmod.mfpackages.sim_tdis_gwf_ims_from_model_ds(model_ds)
+    # create simulation 
+    sim = nlmod.gwf.sim_from_model_ds(model_ds)
+    
+    # create time discretisation
+    tdis = nlmod.gwf.tdis_from_model_ds(model_ds, sim)
+    
+    # create groundwater flow model
+    gwf = nlmod.gwf.gwf_from_model_ds(model_ds, sim)
+    
+    # create ims
+    ims = nlmod.gwf.ims_to_sim(sim)
+    
     # Create discretization
-    nlmod.mfpackages.dis_from_model_ds(model_ds, gwf)
+    nlmod.gwf.dis_from_model_ds(model_ds, gwf)
 
     # create node property flow
-    nlmod.mfpackages.npf_from_model_ds(model_ds, gwf)
+    nlmod.gwf.npf_from_model_ds(model_ds, gwf)
 
     # Create the initial conditions package
-    nlmod.mfpackages.ic_from_model_ds(model_ds, gwf, starting_head=1.0)
+    nlmod.gwf.ic_from_model_ds(model_ds, gwf, starting_head=1.0)
 
     # Create the output control package
-    nlmod.mfpackages.oc_from_model_ds(model_ds, gwf)
+    nlmod.gwf.oc_from_model_ds(model_ds, gwf)
 
     # voeg grote oppervlaktewaterlichamen toe
     da_name = "surface_water"
     model_ds.update(nlmod.read.rws.get_surface_water(model_ds, da_name))
-    nlmod.mfpackages.ghb_from_model_ds(model_ds, gwf, da_name)
+    nlmod.gwf.ghb_from_model_ds(model_ds, gwf, da_name)
 
     # surface level drain
     model_ds.update(nlmod.read.ahn.get_ahn(model_ds))
-    nlmod.mfpackages.surface_drain_from_model_ds(model_ds, gwf)
+    nlmod.gwf.surface_drain_from_model_ds(model_ds, gwf)
 
     # add constant head cells at model boundaries
     model_ds.update(
-        nlmod.mfpackages.constant_head.get_chd_at_model_edge(
+        nlmod.gwf.constant_head.get_chd_at_model_edge(
             model_ds, model_ds["idomain"]
         )
     )
-    nlmod.mfpackages.chd_from_model_ds(model_ds, gwf, head="starting_head")
+    nlmod.gwf.chd_from_model_ds(model_ds, gwf, head="starting_head")
 
     # add knmi recharge to the model datasets
     model_ds.update(nlmod.read.knmi.get_recharge(model_ds))
     # create recharge package
-    nlmod.mfpackages.rch_from_model_ds(model_ds, gwf)
+    nlmod.gwf.rch_from_model_ds(model_ds, gwf)
 
-    nlmod.util.write_and_run_model(gwf, model_ds)
+    nlmod.gwf.write_and_run_model(gwf, model_ds)
 
     return model_ds, gwf
 
@@ -325,43 +355,53 @@ def test_create_sea_model_perlen_14(tmpdir):
     model_ds = model_ds.drop_dims("time")
     model_ds.update(new_model_ds)
 
-    # create modflow packages
-    sim, gwf = nlmod.mfpackages.sim_tdis_gwf_ims_from_model_ds(model_ds)
+    # create simulation 
+    sim = nlmod.gwf.sim_from_model_ds(model_ds)
+    
+    # create time discretisation
+    tdis = nlmod.gwf.tdis_from_model_ds(model_ds, sim)
+    
+    # create groundwater flow model
+    gwf = nlmod.gwf.gwf_from_model_ds(model_ds, sim)
+    
+    # create ims
+    ims = nlmod.gwf.ims_to_sim(sim)
+    
     # Create discretization
-    nlmod.mfpackages.dis_from_model_ds(model_ds, gwf)
+    nlmod.gwf.dis_from_model_ds(model_ds, gwf)
 
     # create node property flow
-    nlmod.mfpackages.npf_from_model_ds(model_ds, gwf)
+    nlmod.gwf.npf_from_model_ds(model_ds, gwf)
 
     # Create the initial conditions package
-    nlmod.mfpackages.ic_from_model_ds(model_ds, gwf, starting_head=1.0)
+    nlmod.gwf.ic_from_model_ds(model_ds, gwf, starting_head=1.0)
 
     # Create the output control package
-    nlmod.mfpackages.oc_from_model_ds(model_ds, gwf)
+    nlmod.gwf.oc_from_model_ds(model_ds, gwf)
 
     # voeg grote oppervlaktewaterlichamen toe
     da_name = "surface_water"
     model_ds.update(nlmod.read.rws.get_surface_water(model_ds, da_name))
-    nlmod.mfpackages.ghb_from_model_ds(model_ds, gwf, da_name)
+    nlmod.gwf.ghb_from_model_ds(model_ds, gwf, da_name)
 
     # surface level drain
     model_ds.update(nlmod.read.ahn.get_ahn(model_ds))
-    nlmod.mfpackages.surface_drain_from_model_ds(model_ds, gwf)
+    nlmod.gwf.surface_drain_from_model_ds(model_ds, gwf)
 
     # add constant head cells at model boundaries
     model_ds.update(
-        nlmod.mfpackages.constant_head.get_chd_at_model_edge(
+        nlmod.gwf.constant_head.get_chd_at_model_edge(
             model_ds, model_ds["idomain"]
         )
     )
-    nlmod.mfpackages.chd_from_model_ds(model_ds, gwf, head="starting_head")
+    nlmod.gwf.chd_from_model_ds(model_ds, gwf, head="starting_head")
 
     # add knmi recharge to the model datasets
     model_ds.update(nlmod.read.knmi.get_recharge(model_ds))
     # create recharge package
-    nlmod.mfpackages.rch_from_model_ds(model_ds, gwf)
+    nlmod.gwf.rch_from_model_ds(model_ds, gwf)
 
-    nlmod.util.write_and_run_model(gwf, model_ds)
+    nlmod.gwf.write_and_run_model(gwf, model_ds)
 
     return model_ds, gwf
